+++ src/packages/components.ts
@@ -9,3 +9,5 @@ export * from './input-number';
 export * from './checkbox';
 export * from './radio';
 export * from './switch';
+export * from './select-down';
+export * from "./tag";